PCapConversion/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
scapy>=2.5.0
requests>=2.28.0
azure-identity>=1.12.0
azure-monitor-ingestion>=1.0.0
